Remove BinaryPopulation and AbstractPopulation
