+++ gp/cgp_population.py
@@ -1,13 +1,15 @@
-from .abstract_population import AbstractPopulation
+import numpy as np
+
 from .cgp_individual import CGPIndividual, CGPIndividualMultiGenome
 
 
-class CGPPopulation(AbstractPopulation):
+class CGPPopulation:
+    """
+    A population of individuals.
+    """
+
     def __init__(self, n_parents, mutation_rate, seed, genome_params):
         """Init function.
-
-        Extends AbstractPopulation.__init__
-
 
         Parameters
         ----------
@@ -21,9 +23,52 @@ class CGPPopulation(AbstractPopulation):
         genome_params : dict
             Parameters for the genomes of the population's individuals.
         """
+        self.n_parents = n_parents  # number of individuals in parent population
+
+        if not (0.0 < mutation_rate and mutation_rate < 1.0):
+            raise ValueError("mutation rate needs to be in (0, 1)")
+        self._mutation_rate = mutation_rate  # probability of mutation per gene
+
+        self.seed = seed
+        self.rng = np.random.RandomState(self.seed)
+
         self._genome_params = genome_params
 
-        super().__init__(n_parents, mutation_rate, seed)
+        self._parents = None  # list of parent individuals
+
+        # keeps track of the number of generations, increases with
+        # every new offspring generation
+        self.generation = 0
+        self._max_idx = -1  # keeps track of maximal idx in population used to label individuals
+
+        self._generate_random_parent_population()
+
+    @property
+    def champion(self):
+        """Return parent with the highest fitness.
+        """
+        return max(self._parents, key=lambda ind: ind.fitness)
+
+    @property
+    def parents(self):
+        return self._parents
+
+    @parents.setter
+    def parents(self, new_parents):
+        self.generation += 1
+        self._parents = new_parents
+
+    def __getitem__(self, idx):
+        return self._parents[idx]
+
+    def _generate_random_parent_population(self):
+        self._parents = self._generate_random_individuals(self.n_parents)
+        self._label_new_individuals(self._parents)
+
+    def _label_new_individuals(self, individuals):
+        for ind in individuals:
+            ind.idx = self._max_idx + 1
+            self._max_idx += 1
 
     def _generate_random_individuals(self, n):
         individuals = []
@@ -42,7 +87,78 @@ class CGPPopulation(AbstractPopulation):
         return individuals
 
     def crossover(self, breeding_pool, n_offsprings):
+        """Create an offspring population via crossover.
+
+        Parameters
+        ----------
+        breeding_pool : List[gp.CGPPopulation]
+            List of individuals from which the offspring are created.
+        n_offsprings : int
+            Number of offspring to be created.
+
+        Returns
+        ----------
+        List[gp.AbstractIndividual]
+            List of offspring individuals.
+        """
+        # in principle crossover would rely on a procedure like the
+        # following:
+        # offsprings = []
+        # while len(offsprings) < n_offsprings:
+        #     first_parent, second_parent = self.rng.permutation(breeding_pool)[:2]
+        #     offsprings.append(first_parent.crossover(second_parent, self.rng))
+
+        # return offsprings
+        # however, as cross over tends to disrupt the search in CGP
+        # (Miller, 1999) crossover is skipped, instead the best
+        # individuals from breeding pool are returned.
+        # reference:
+        # Miller, J. F. (1999). An empirical study of the efficiency
+        # of learning boolean functions using a cartesian genetic
+        # programming approach. In Proceedings of the 1st Annual
+        # Conference on Genetic and Evolutionary Computation-Volume 2,
+        # pages 1135–1142. Morgan Kaufmann Publishers Inc.
         assert len(breeding_pool) >= n_offsprings
-        # do not perform crossover for CGP, just choose the best
-        # individuals from breeding pool
         return sorted(breeding_pool, key=lambda x: -x.fitness)[:n_offsprings]
+
+    def mutate(self, offsprings):
+        """Mutate a list of offspring invididuals.
+
+        Parameters
+        ----------
+        offsprings : List[gp.AbstractIndividual]
+            List of offspring individuals to be mutated.
+
+        Returns
+        ----------
+        List[gp.AbstractIndividual]
+            List of mutated offspring individuals.
+        """
+
+        for off in offsprings:
+            off.mutate(self._mutation_rate, self.rng)
+        return offsprings
+
+    def fitness_parents(self):
+        """Return fitness for all parents of the population.
+
+        Returns
+        ----------
+        List[float]
+            List of fitness values for all parents.
+        """
+        return [ind.fitness for ind in self._parents]
+
+    def dna_parents(self):
+        """Return a list of the DNA of all parents.
+
+        Returns
+        ----------
+        List[List[int]]
+            List of dna of all parents.
+        """
+
+        dnas = np.empty((self.n_parents, self._parents[0].genome._n_genes))
+        for i in range(self.n_parents):
+            dnas[i] = self._parents[i].genome.dna
+        return dnas
