+++ test/test_ea_mu_plus_lambda.py
@@ -5,6 +5,21 @@ import gp
 
 SEED = np.random.randint(2 ** 31)
 
+population_params = {
+    "n_parents": 5,
+    "mutation_rate": 0.05,
+    "seed": SEED,
+}
+
+genome_params = {
+    "n_inputs": 2,
+    "n_outputs": 1,
+    "n_columns": 3,
+    "n_rows": 3,
+    "levels_back": 2,
+    "primitives": [gp.CGPAdd, gp.CGPSub, gp.CGPMul, gp.CGPConstantFloat],
+}
+
 
 def test_label():
     def objective_without_label(individual):
@@ -17,7 +32,8 @@ def test_label():
         individual.fitness = -1
         return individual
 
-    pop = gp.BinaryPopulation(1, 0.5, SEED, {"genome_length": 2, "primitives": [0, 1]})
+    pop = gp.CGPPopulation(**population_params, genome_params=genome_params)
+
     ea = gp.ea.MuPlusLambda(1, 2, 1)
     ea.initialize_fitness_parents(pop, objective_without_label)
     ea.step(pop, objective_without_label)
@@ -32,7 +48,8 @@ def test_fitness_contains_nan():
             individual.fitness = np.random.rand()
         return individual
 
-    pop = gp.BinaryPopulation(5, 0.5, SEED, {"genome_length": 2, "primitives": [0, 1]})
+    pop = gp.CGPPopulation(**population_params, genome_params=genome_params)
+
     ea = gp.ea.MuPlusLambda(10, 10, 1)
     ea.initialize_fitness_parents(pop, objective)
     ea.step(pop, objective)
